Move command to get undercloud IP to answers file
Since more modern distributions of libvirt, such as the one included in
Fedora 23, no longer use the default.leases file. It's useful to move
the command that gets the undercloud's IP to the answers file, so it can
be edited if one uses a distribution that needs another method of doing
so.