Get undercloud IP using virsh if no default.leases exists
Since the usage of the default.leases file was deprecated from libvirt,
there is a need to have a the getting of the undercloud IP in such a way
that it works in both the new and the old way. So this commit addresses
that.

On the other hand, with the new method, for some reason it will not get
the IP if the stack user is used with sudo, in ansible. So this was
removed.